Reafactor into new datasources module
