+++ datasources/src/main/java/org/opensearch/sql/datasources/model/UpdateDataSourceActionResponse.java
@@ -1,0 +1,33 @@
+/*
+ *
+ *  * Copyright OpenSearch Contributors
+ *  * SPDX-License-Identifier: Apache-2.0
+ *
+ */
+
+package org.opensearch.sql.datasources.model;
+
+import java.io.IOException;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.opensearch.action.ActionResponse;
+import org.opensearch.common.io.stream.StreamInput;
+import org.opensearch.common.io.stream.StreamOutput;
+
+@RequiredArgsConstructor
+public class UpdateDataSourceActionResponse
+    extends ActionResponse {
+
+  @Getter
+  private final String result;
+
+  public UpdateDataSourceActionResponse(StreamInput in) throws IOException {
+    super(in);
+    result = in.readString();
+  }
+
+  @Override
+  public void writeTo(StreamOutput streamOutput) throws IOException {
+    streamOutput.writeString(result);
+  }
+}
