+++ plugin/build.gradle
@@ -122,14 +122,12 @@ dependencies {
     api "com.fasterxml.jackson.core:jackson-core:${versions.jackson}"
     api "com.fasterxml.jackson.core:jackson-databind:${versions.jackson_databind}"
     api "com.fasterxml.jackson.core:jackson-annotations:${versions.jackson}"
-    api group: 'commons-io', name: 'commons-io', version: '2.8.0'
-    implementation group: 'org.opensearch', name: 'opensearch-x-content', version: "${opensearch_version}"
-    implementation group: 'org.opensearch', name: 'common-utils', version: "${opensearch_build}"
 
     api project(":ppl")
     api project(':legacy')
     api project(':opensearch')
     api project(':prometheus')
+    api project(':datasources')
 
     testImplementation group: 'net.bytebuddy', name: 'byte-buddy-agent', version: '1.12.13'
     testImplementation group: 'org.hamcrest', name: 'hamcrest-library', version: '2.1'
