+++ plugin/src/main/java/org/opensearch/sql/plugin/SQLPlugin.java
@@ -46,13 +46,15 @@ import org.opensearch.rest.RestHandler;
 import org.opensearch.script.ScriptContext;
 import org.opensearch.script.ScriptEngine;
 import org.opensearch.script.ScriptService;
-import org.opensearch.sql.common.encryptor.EncryptorImpl;
-import org.opensearch.sql.datasource.DataSourceLoaderCache;
+import org.opensearch.sql.datasources.encryptor.EncryptorImpl;
 import org.opensearch.sql.datasource.DataSourceMetadataStorage;
 import org.opensearch.sql.datasource.DataSourceService;
 import org.opensearch.sql.datasource.DataSourceServiceImpl;
 import org.opensearch.sql.datasource.DataSourceUserAuthorizationHelper;
-import org.opensearch.sql.datasource.model.DataSource;
+import org.opensearch.sql.datasources.transport.TransportCreateDataSourceAction;
+import org.opensearch.sql.datasources.transport.TransportDeleteDataSourceAction;
+import org.opensearch.sql.datasources.transport.TransportGetDataSourceAction;
+import org.opensearch.sql.datasources.transport.TransportUpdateDataSourceAction;
 import org.opensearch.sql.legacy.esdomain.LocalClusterState;
 import org.opensearch.sql.legacy.executor.AsyncRestExecutor;
 import org.opensearch.sql.legacy.metrics.Metrics;
@@ -65,24 +67,20 @@ import org.opensearch.sql.opensearch.storage.OpenSearchDataSourceFactory;
 import org.opensearch.sql.opensearch.storage.script.ExpressionScriptEngine;
 import org.opensearch.sql.opensearch.storage.serialization.DefaultExpressionSerializer;
 import org.opensearch.sql.plugin.config.OpenSearchPluginModule;
-import org.opensearch.sql.plugin.datasource.DataSourceSettings;
-import org.opensearch.sql.plugin.datasource.DataSourceUserAuthorizationHelperImpl;
-import org.opensearch.sql.plugin.datasource.OpenSearchDataSourceMetadataStorage;
-import org.opensearch.sql.plugin.model.CreateDataSourceActionResponse;
-import org.opensearch.sql.plugin.model.DeleteDataSourceActionResponse;
-import org.opensearch.sql.plugin.model.GetDataSourceActionResponse;
-import org.opensearch.sql.plugin.model.UpdateDataSourceActionResponse;
-import org.opensearch.sql.plugin.rest.RestDataSourceQueryAction;
+import org.opensearch.sql.datasources.settings.DataSourceSettings;
+import org.opensearch.sql.datasources.auth.DataSourceUserAuthorizationHelperImpl;
+import org.opensearch.sql.datasources.storage.OpenSearchDataSourceMetadataStorage;
+import org.opensearch.sql.datasources.model.CreateDataSourceActionResponse;
+import org.opensearch.sql.datasources.model.DeleteDataSourceActionResponse;
+import org.opensearch.sql.datasources.model.GetDataSourceActionResponse;
+import org.opensearch.sql.datasources.model.UpdateDataSourceActionResponse;
+import org.opensearch.sql.datasources.rest.RestDataSourceQueryAction;
 import org.opensearch.sql.plugin.rest.RestPPLQueryAction;
 import org.opensearch.sql.plugin.rest.RestPPLStatsAction;
 import org.opensearch.sql.plugin.rest.RestQuerySettingsAction;
 import org.opensearch.sql.plugin.transport.PPLQueryAction;
 import org.opensearch.sql.plugin.transport.TransportPPLQueryAction;
 import org.opensearch.sql.plugin.transport.TransportPPLQueryResponse;
-import org.opensearch.sql.plugin.transport.datasource.TransportCreateDataSourceAction;
-import org.opensearch.sql.plugin.transport.datasource.TransportDeleteDataSourceAction;
-import org.opensearch.sql.plugin.transport.datasource.TransportGetDataSourceAction;
-import org.opensearch.sql.plugin.transport.datasource.TransportUpdateDataSourceAction;
 import org.opensearch.sql.prometheus.storage.PrometheusStorageFactory;
 import org.opensearch.sql.storage.DataSourceFactory;
 import org.opensearch.threadpool.ExecutorBuilder;
