--- plugin/src/main/java/org/opensearch/sql/plugin/model/DeleteDataSourceActionResponse.java
@@ -1,33 +1,0 @@
-/*
- *
- *  * Copyright OpenSearch Contributors
- *  * SPDX-License-Identifier: Apache-2.0
- *
- */
-
-package org.opensearch.sql.plugin.model;
-
-import java.io.IOException;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.opensearch.action.ActionResponse;
-import org.opensearch.common.io.stream.StreamInput;
-import org.opensearch.common.io.stream.StreamOutput;
-
-@RequiredArgsConstructor
-public class DeleteDataSourceActionResponse extends ActionResponse {
-
-  @Getter
-  private final String result;
-
-  public DeleteDataSourceActionResponse(StreamInput in) throws IOException {
-    super(in);
-    result = in.readString();
-  }
-
-  @Override
-  public void writeTo(StreamOutput streamOutput) throws IOException {
-    streamOutput.writeString(result);
-  }
-
-}
