+++ settings.gradle
@@ -19,3 +19,5 @@ include 'legacy'
 include 'sql'
 include 'prometheus'
 include 'benchmarks'
+include 'datasources'
+
